server/requirements.txt
Django==3.2.25
django-environ==0.8.1
django-cors-headers==3.12.0
django_extensions==3.1.5
djangorestframework==3.13.1
djangorestframework-simplejwt==5.2.0
psycopg2-binary==2.9.2
gunicorn==22.0.0